+++ src/hittable_list.rs
@@ -17,11 +17,13 @@ impl HittableList {
 }
 
 impl Hittable for HittableList {
-    fn hit(&self, r: &Ray, t_min: f32, t_max: f32, rec: &mut HitRecord) -> bool {
+    fn hit(&self, r: &Ray, t_min: f64, t_max: f64, rec: &mut HitRecord) -> bool {
         let mut temp_rec: HitRecord = HitRecord::default();
         let mut hit_anything: bool = false;
-        let mut closest_so_far: f32 = t_max;
+        let mut closest_so_far: f64 = t_max;
 
+        // iterate over all hittables and call hit
+        // Return the closest hit
         for object in &self.objects {
             if object.hit(r, t_min, closest_so_far, &mut temp_rec) {
                 hit_anything = true;
